Correct bug

+++ libControl/FFmpegTranscodingPimpl.h
@@ -138,5 +138,5 @@ private:
 	SwsContext* scaleContext = nullptr;
 	bool m_allowSeek = true;
 	int videoStreamIndex = 0;
-	float startTime = 0;
+	int64_t startTime = 0;
 };
